src/test/java/com/example/stuffmonolith/StuffControllerTests.java
package com.example.stuffmonolith;

import com.fasterxml.jackson.databind.ObjectMapper;
import lombok.AllArgsConstructor;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.test.web.servlet.MockMvc;

import java.util.List;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@WebMvcTest(StuffController.class)
@AllArgsConstructor
public class StuffControllerTests {
    @Autowired
    private MockMvc mockMvc;

    @Autowired
    private ObjectMapper objectMapper;

    @Test
    void shouldReturnLampSweatshirtKakapo() throws Exception {
        var expectedStuff = List.of(
                new Item(1L, "lamp"),
                new Item(2L, "sweatshirt"),
                new Item(3L, "kakapo"));

        mockMvc.perform(get("/stuff"))
                .andExpect(status().isOk())
                .andExpect(content().json(objectMapper.writeValueAsString(expectedStuff)));
    }
}

